restructured the project to have each interface in one folder.
