+++ src/Microsoft.PowerShell.IoT/Microsoft.PowerShell.IoT.csproj
@@ -16,6 +16,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="System.Device.Gpio" Version="1.0.0" />
     <PackageReference Include="System.Management.Automation" Version="6.0.0-rc.2" />
     <PackageReference Include="Unosquare.Raspberry.IO" Version="0.14.0" />
   </ItemGroup>
